pyproject.toml
[project]
name = "mlops-colorflow"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "datasets==4.0.0",
    "diagrams>=0.25.1",
    "dvc[gs]>=3.67.1",
    "fastai>=2.7",
    "gcloud>=0.18.3",
    "hydra-core>=1.3",
    "ipython>=9.13.0",
    "matplotlib==3.10.0",
    "mlflow>=2.10",
    "numpy>=1.26",
    "pillow>=10.0.0",
    "scikit-image==0.26.0",
    "setuptools==69.5.1", # TODO: Remove if not needed anymore
    "tensorboard==2.20.0", # TODO: Remove after optimization
    "tensorboard-data-server==0.7.2", # TODO: Remove after optimization
    "torch>=2.2",
    "torchvision>=0.17",
    "tqdm>=4.66",
]

[project.optional-dependencies]
# Optionally add hydra-optuna-sweeper later via:
#   uv add --prerelease=allow hydra-optuna-sweeper
hpo = [
    "optuna>=3.5",
]
ray = [
    "ray[tune]>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/colorflow"]
